refactor(Client): Move event listeners to documents file
